+++ evaluation/nginx/second_run_gear_images_without_cache.py
@@ -53,8 +53,9 @@ class Runner:
             for tag in tags:
                 private_repo = private_registry + repo + suffix + ":" + tag
 
-                if os.path.exists(localVolume) == False:
-                    os.makedirs(localVolume)
+                if localVolume != "":
+                    if os.path.exists(localVolume) == False:
+                        os.makedirs(localVolume)
 
                 print "start running: ", private_repo
 
@@ -132,7 +133,8 @@ class Runner:
                 else:
                     time.sleep(5)
 
-                shutil.rmtree(localVolume)
+                if localVolume != "":
+                    shutil.rmtree(localVolume)
 
 
     def record(self, repo, tag, time):
